Prevent Ink from altering the accessibility tree
When rendered inside a `<button>`, canvas element produces an unintended side-effect in Chrome, changing the announced role of the button to "button, group". Adding `aria-hidden="true"` effectively hides the canvas element from assistive software such as screen readers, so the role of the button is correctly announced as just "button".